Fixes webcamDaemon for Logitech C170
For whatever reason using the -f parameter with that specific camera
model causes it to just not output any image. mjpgstreamer will start
up just fine and not log an error or anything, the camera will appear
to start up too, but when connecting to the webcam stream nothing will
be displayed at all. Just removing the -f parameter in this case fixes
it completely.

This patch adds a new check to the webcamDaemon when starting up a
USB webcam which checks the webcams USB vid and pid against a list
of webcams known not to work with -f (right now that's just the C170,
but more can be added easily by the user through the new
additional_brokenfps_usb_devices option in /boot/octopi.txt) and if
a match is found strips the -f parameter from the camera options.

This is the only way I found I could make a C170 work. It would
probably be cleaner to try to debug what's going wrong in mjpgstreamer
there, but since that might as well be a problem with the chipset
driver of the C170, this is a workaround that should at least
yield a picture at a default frame rate.