Added ability to resize root partition
If OCTOPI_IMAGE_ENLARGEROOT is set to a size (in MB), the
root partition of the raspbian image will be enlarged by that
size before starting the actual build. This helps in situations
where the image doesn't ship with enough space to perform the build.

If OCTOPI_IMAGE_RESIZEROOT is set to a size (in MB), after the
build the image will be resized to its minimal size plus the provided
size. 200 MB appears to be a good value here.