Speed up preparation of oprint environment
We fetch the raspbian packages and use the precompiled versions
of the pyopenssl related libraries instead of having the build
spend what feels like hours on building the cryptography module.

We also no longer need to manually install pyserial, the version
on pypi is fixed and OctoPrint will pull that in automatically.