Added install-desktop script
Allows installing the desktop environment which is usually not
installed on raspbian-lite.

Some users might require X (although it is not really recommended
to run something as heavy weight as that next to OctoPrint on a Pi,
at least not on single core Pi's, but it will still be easier to
help those users if there's a simple script they can execute than
having to explain the necessary steps.