Change password of user pi and hostname via /boot
By creating a file /boot/octopi-password.txt and/or a file
/boot/octopi-hostname.txt, the user can now change the password
of the user pi and the hostname of the whole instance without
having to boot the pi first.

The system will detect that such files exist, read them, apply
their contents as new password/hostname and then delete the file
if that was successful. Hostnames are checked if they are valid
hostnames (only alphanumeric characters and hyphen are allowed).
Passwords are checked if they are empty (no further restrictions
according to the docs). Both are left and right trimmed.

This allows a completely headless and even ssh-less initial setup
of OctoPi, just by creating/editing a couple of .txt files on the
FAT partition that is used for /boot, and hence can be done easily
alongside flashing.