Add "how to use" instructions to boot screen
A lot of people appear to get confused when they boot
up OctoPi with a screen attached and are then greeted
by a command line login instead of OctoPrint.

Therefore added instructions to please use a browser
to connect to a list of potential addresses to access
OctoPrint - that might help.