Remove OctoPiPanel and display scripts
OctoPiPanel is no longer actively maintained according to its
page and the included display scripts are horribly outdated
and probably broken beyond repair.

Considering the amount of RPi displays at this point, it's
probably sensible to rather instruct users to refer to vendor
instructions on how to setup a specific display against a
Raspbian (lite) image instead of trying to accomodate that
ever changing landscape.