Move gencert from SysVInit to systemd
For some reason, no matter what I tried the old SysVInit style
init script was no longer starting on boot under Stretch based
images. That meant that no cert was generated on initial boot
and haproxy could not start due to the missing SSL cert as
also mentioned in #402

Switching it to a simple systemd oneshot service solved the
issue.