Install vcgencmd and fix user permission on Ubuntu
The 64bit images are currently lacking vcgencmd, meaning that the
PiSupport plugin can't check for undervoltage and overheat
situations and will complain about this in the latest version
as well, see feedback in #770.

This should fix it by installing the package containing vcgencmd
and making sure the base user is added to the video group as
well.

I was not able to test this in a build since I could not find
documentation on which base image exactly to use and how
to precisely run the 64bit build, but based on user feedback
it should hopefully work. Still, please test in a build before
merging.